Iso camera. Movement. Shooting. Walk up slopes.
Oh baby, this commit's got it all!

I got angry and deleted everything in the last commit.
I've been following this tutorial:
https://www.youtube.com/playlist?list=PLFt_AvWsXl0ct75VmGXskJnzrStEbPD6d
I got motion. You can choose (in the code, NOT in play) whether to
control the duder with WASD, or WASD+Mouse. WASD+Mouse is a bit buggy,
he wobbles some...
NO jumping yet.
YES he can successfully walk up slopes. Looks like 45° is his limit.
Fucked if I know how it decides that.

He can shoot. With his cube gun. It plays a sound, and shows a flash of
light for a frame. There's a variable, GunType, you can change to be
Automatic or not (Hold down mouse to autofire). The sound gets weird
when firing fast.